backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
mssql-django
pyodbc
python-dotenv
pandas
xlsxwriter